# api_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusConnect API",
    "description": "B2B SaaS platform for institutions and scholarships",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scholarships": {
      "get": {
        "tags": [
          "scholarships"
        ],
        "summary": "Get Scholarships",
        "description": "Get all scholarships with optional filters",
        "operationId": "get_scholarships_api_v1_scholarships_get",
        "parameters": [
          {
            "name": "scholarship_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scholarship Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScholarshipResponse"
                  },
                  "title": "Response Get Scholarships Api V1 Scholarships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scholarships/{scholarship_id}": {
      "get": {
        "tags": [
          "scholarships"
        ],
        "summary": "Get Scholarship",
        "description": "Get a single scholarship by ID",
        "operationId": "get_scholarship_api_v1_scholarships__scholarship_id__get",
        "parameters": [
          {
            "name": "scholarship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scholarship Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScholarshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/subscriptions/create-checkout": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a Stripe checkout session for premium subscription",
        "operationId": "create_checkout_session_api_v1_admin_subscriptions_create_checkout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/subscriptions/current": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Get Current Subscription",
        "description": "Get current subscription status",
        "operationId": "get_current_subscription_api_v1_admin_subscriptions_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/subscriptions/cancel": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Cancel Subscription",
        "description": "Cancel subscription at end of period",
        "operationId": "cancel_subscription_api_v1_admin_subscriptions_cancel_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/subscriptions/portal": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Get Customer Portal",
        "description": "Get Stripe customer portal URL",
        "operationId": "get_customer_portal_api_v1_admin_subscriptions_portal_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScholarshipResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "organization": {
            "type": "string",
            "title": "Organization"
          },
          "scholarship_type": {
            "type": "string",
            "title": "Scholarship Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "difficulty_level": {
            "type": "string",
            "title": "Difficulty Level"
          },
          "amount_min": {
            "type": "integer",
            "title": "Amount Min"
          },
          "amount_max": {
            "type": "integer",
            "title": "Amount Max"
          },
          "is_renewable": {
            "type": "boolean",
            "title": "Is Renewable"
          },
          "number_of_awards": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Awards"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "application_opens": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Opens"
          },
          "for_academic_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "For Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "website_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website Url"
          },
          "min_gpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Gpa"
          },
          "primary_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Image Url"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "views_count": {
            "type": "integer",
            "title": "Views Count"
          },
          "applications_count": {
            "type": "integer",
            "title": "Applications Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "organization",
          "scholarship_type",
          "status",
          "difficulty_level",
          "amount_min",
          "amount_max",
          "is_renewable",
          "verified",
          "featured",
          "id",
          "views_count",
          "applications_count",
          "created_at"
        ],
        "title": "ScholarshipResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/admin/auth/login"
          }
        }
      }
    }
  }
}